backend/requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.4
daphne==4.0.0
Django==5.0
django-cors-headers==4.3.1
django-fcm==0.1.1
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
firebase-admin==6.5.0
frozenlist==1.4.0
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.7
multidict==6.0.4
phonenumbers==8.13.26
Pillow==10.1.0
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
pyparsing==3.1.2
python-memcached==1.62
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
rsa==4.9
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
twilio==8.10.3
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.4
zope.interface==6.2
